First pass at propagating info via headers for gql spans
